+++ lib/src/utils.ts
@@ -4,12 +4,15 @@
  * @packageDocumentation
  */
 export * from './utils/arrayBufferToString.js';
+export * from './utils/arrayBufferToUuid.js';
 export * from './utils/base64decode.js';
 export * from './utils/base64encode.js';
 export * from './utils/convertUint8ToUint16.js';
 export * from './utils/dataViewToString.js';
 export * from './utils/decodeBase64.js';
+export * from './utils/decodeHex.js';
 export * from './utils/encodeBase64.js';
+export * from './utils/encodeHex.js';
 export * from './utils/Encoding.js';
 export * from './utils/roundToEven.js';
 export * from './utils/stringToUint16.js';
@@ -20,5 +23,6 @@ export * from './utils/utf8ArrayToStr.js';
 export * from './utils/UTF_16.js';
 export * from './utils/UTF_8.js';
 export * from './utils/uuid.js';
+export * from './utils/uuidToArrayBuffer.js';
 export type { ValueOf } from './utils/ValueOf.js';
 export type { ValueOrArray } from './utils/ValueOrArray.js';