+++ lib/test/webvtt/parseWebVtt.test.ts
@@ -8,7 +8,7 @@ describe('parseWebVtt examples', () => {
 		//#region example
 		const vtt = `WEBVTT\n\nREGION\nid:test\n\nSTYLE\n::cue {}\n\nCUE_1\n00:00:00.000 --> 00:00:35.000\nWevVTT Sample\n`;
 
-		const { cues, errors, regions, styles } = parseWebVtt(vtt);
+		const { cues, errors, regions, styles } = await parseWebVtt(vtt);
 
 		assert(cues[0].id === 'CUE_1');
 		assert(cues[0].text === 'WevVTT Sample');