+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to
 
 - Add CMCD config type and request builder
   ([#227](https://github.com/streaming-video-technology-alliance/common-media-library/issues/227))
+- Add UUID utilities for converting to and from hex strings and changing
+  endianness
 
 ## [0.16.0] - 2025-07-09