+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to
 
 - Add CMCD config type and request builder
   ([#227](https://github.com/streaming-video-technology-alliance/common-media-library/issues/227))
+- Add const for DASH thumbnail tile scheme ID URI
 
 ### Changed